requirements.txt
boto3==1.18.11
botocore==1.21.11
configparser==5.0.2
jmespath==0.10.0
numpy==1.21.1
pandas==1.3.1
psycopg2-binary==2.9.1
pyathena==2.3.0
python-dateutil==2.8.2
pytz==2021.1
s3transfer==0.5.0
six==1.16.0
tenacity==8.0.1
urllib3==1.26.6
